server/requirement.txt
autopep8==2.0.2
cassandra-driver==3.25.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-CQLAlchemy==2.0.0
geomet==0.2.1.post1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.3
